ai-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.26.2
rank-bm25==0.2.2
python-dotenv==1.0.0
